Async Texture loading now working. The very slow loading screen is for demonstration purpos.
If no texture was loaded before the texture loading is syncroniced. Else it is done asyncroniced and evey object without a loaded texture uses the first loaded texture.
The Texture Loader thread is forced to only load one texture every 10 seconds to demonstrate that the main thread can render without it.